app/backend/requirements.txt
flask~=1.1.0
flask-restx~=0.2.0
PyYAML~=5.3.0
yamale~=3.0.0
google-api-python-client~=1.10.0
pygit2~=1.3.0
python-json-logger~=2.0.0
python-gitlab~=2.5.0
google-auth~=1.22.0
google-cloud-storage~=1.31.0
waitress~=1.4.0
google-api-core~=1.23.0
grpcio~=1.33.1
oauth2client~=4.1.0
google-cloud-secret-manager~=2.0.0
cryptography~=3.2.0
google-cloud-iam~=2.0.0

